+++ src/utils.ts
@@ -1,6 +1,10 @@
 import * as vscode from "vscode";
 import * as fs from 'fs';
+import * as path from 'path';
+import cp = require('child_process');
 import { TextSelection, ABLTableDefinition, ABLIndexDefinition } from "./definition";
+import { getConfig } from "./ablConfig";
+import { getXcodeBin } from "./environment";
 
 let regexInvalidWordEnd: RegExp = new RegExp(/[\.|\:|\-|\_|\\|\/]$/);
 
@@ -178,4 +182,44 @@ export function saveAndExec(document: vscode.TextDocument, action: Function) {
 	}
 	else
 		action();
+}
+
+export function xcode(filename: string): Promise<boolean> {
+	let baseName = path.basename(filename);
+	let tmpDir = vscode.workspace.rootPath;
+	// temp xcode dir
+	tmpDir = [tmpDir,'.xc'+Math.floor(Math.random()*100000).toString()].join('\\');
+	// temp xcode output
+	let outDir = [tmpDir, 'out'].join('\\');
+	// mkdir (outDir is inside tmpDir)
+	mkdir(outDir);
+	// copy file to temp dir
+	let oldFilename = [vscode.workspace.rootPath, baseName].join('\\');
+	let newFilename = [tmpDir, baseName].join('\\');
+	fs.copyFileSync(oldFilename, newFilename);
+	// exec xcode
+	let cwd = path.dirname(newFilename);
+	let cmd = getXcodeBin();
+
+	let args = [
+		'-d',
+		outDir,
+		baseName
+	];
+	return new Promise<boolean>((resolve, reject) => {
+		cp.execFile(cmd, args, { cwd: cwd }, (err, stdout, stderr) => {
+			// copy xcoded file to overwrite original file
+			fs.copyFileSync([outDir,baseName].join('\\'), oldFilename);
+			// remove files/dorectories
+			fs.unlinkSync([outDir,baseName].join('\\'));
+			fs.unlinkSync(newFilename);
+			fs.rmdirSync(outDir);
+			fs.rmdirSync(tmpDir);
+			//
+			if (err) 
+				resolve(false);
+			else
+				resolve(true);
+		});
+	});
 }